api/requirements.txt
fastapi
uvicorn
python-multipart
pure-magic
ExifRead
pdfminer.six
python-docx
beautifulsoup4
hachoir
# yara-python # Commenting out as it may fail on Vercel without C libs
